+++ src/js/Grid.js
@@ -80,7 +80,6 @@ var Grid = /** @class */ (function () {
             }
             else {
                 gf = Math.abs(Math.floor(Grid.utils.baseLog(1 / adjustFactor, 1 / (scale + 1))));
-                //gf = Math.pow( adjustFactor, gf );
             }
             return gf;
         }